Move fonts handling to the utils package
