+++ weblate/trans/widgets.py
@@ -32,7 +32,7 @@ from django.template.loader import render_to_string
 
 from PIL import Image, ImageDraw
 
-from weblate.trans.fonts import is_base, get_font
+from weblate.utils.fonts import is_base, get_font
 from weblate.utils.site import get_site_url
