+++ weblate/utils/fonts.py
